chore: remove redundancies
