--- packages/frontend/src/components/Explorer/Explorer.styles.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const ExplorerWrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    flex: 1;
-`;
