+++ packages/frontend/src/components/common/Page/Sidebar.tsx
@@ -1,27 +1,19 @@
 import { Card, Divider, H3 } from '@blueprintjs/core';
 import React, { FC } from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { TrackSection } from '../../../providers/TrackingProvider';
 import { SectionName } from '../../../types/Events';
 import AboutFooter from '../../AboutFooter';
 
-const SidebarWrapper = styled(Card)<{ noMargin?: boolean }>`
+const SidebarWrapper = styled(Card)`
     height: calc(100vh - 50px);
     flex-basis: 400px;
     flex-shrink: 0;
     flex-grow: 0;
-    margin-right: 10px;
     overflow: hidden;
     position: sticky;
     top: 50px;
     padding-bottom: 0;
-
-    ${({ noMargin }) =>
-        noMargin &&
-        css`
-            margin-right: 0;
-            z-index: 1;
-        `}
 `;
 
 const SidebarColumn = styled('div')`
@@ -40,15 +32,14 @@ const SidebarContent = styled('div')`
 
 interface SidebarProps {
     title?: string | React.ReactNode;
-    noMargin?: boolean;
 }
 
 export const SidebarDivider = styled(Divider)`
     margin: 12px 0 18px 0;
 `;
 
-const Sidebar: FC<SidebarProps> = ({ title, noMargin, children }) => (
-    <SidebarWrapper elevation={1} noMargin={noMargin}>
+const Sidebar: FC<SidebarProps> = ({ title, children }) => (
+    <SidebarWrapper elevation={1}>
         <TrackSection name={SectionName.SIDEBAR}>
             <SidebarColumn>
                 <SidebarContent>
